# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Project Management System
  version: 1.0.0
  description: >
    A RESTful API for managing projects, tasks, and comments with role-based access control.
    All protected endpoints require a Bearer JWT token obtained from `/auth/login`.
servers:
  - url: 'http://localhost:3000'
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Comment:
      type: object
      properties:
        _id:
          type: string
          example: 6a1df2921ce3c86113503399
        content:
          type: string
          example: This looks good, merging now.
        taskId:
          type: string
          example: 6a1dc57109714d00e4dd6ee1
        userId:
          type: string
          example: 6a0eec205a0dabe0d64ad30a
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
 
    # ─── Validation Error ────────────────────────────────────
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        statusCode:
          type: integer
          example: 400
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ['Email must be a valid email address']
            password: ['Password must be at least 8 characters long']
    UnauthorizedError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Unauthorized
        statusCode:
          type: integer
          example: 401
    ForbiddenError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 
            Access denied: You are not authorized to perform this action
        statusCode:
          type: integer
          example: 403
    NotFoundError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Resource not found
        statusCode:
          type: integer
          example: 404
  # ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /tasks/{taskId}/comments:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
        example: 6a1dc57109714d00e4dd6ee1
 
    post:
      tags: [Comments]
      summary: Create a comment on a task
      description: Adds a new comment authored by the authenticated user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 3
                  maxLength: 200
                  example: Looks good, merging now.
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: Comment created successfully
                  comment:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
 
    get:
      tags: [Comments]
      summary: Get all comments for a task
      description: Returns every comment associated with the given task.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Comments fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Comments fetched successfully
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /tasks/{taskId}/comments/{id}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
        example: 6a1dc57109714d00e4dd6ee1
      - name: id
        in: path
        required: true
        schema:
          type: string
        example: 6a1df2921ce3c86113503399
 
    get:
      tags: [Comments]
      summary: Get a single comment
      description: Returns details of one comment by its ID.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Comment fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Comment fetched successfully
                  comment:
                    $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Comment or task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
 
    patch:
      tags: [Comments]
      summary: Update a comment
      description: >
        Updates the content of an existing comment.
        Only the comment author may do this (`CheckOwnerShip('Comment')` guard).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 3
                  maxLength: 200
                  example: Updated comment content here.
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Comment updated successfully
                  comment:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden – not the comment author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Comment or task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
 
    delete:
      tags: [Comments]
      summary: Delete a comment
      description: >
        Permanently deletes the comment.
        Only the comment author may do this (`CheckOwnerShip('Comment')` guard).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Comment deleted successfully
                  data:
                    type: object
                    nullable: true
                    example: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden – not the comment author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Comment or task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'